Add files via upload
The codes in these files have been modified (some files being added) from the actual code ParTI by Uri Alon group. The code was modified for representing the datapoint in the plot with different colors based on the cell type the each datapoint belongs to, and to store the dimensionality reduced PCA points as an additional CSV file for further analysis.